Refactor Copy to Clipboard Logic
